CHANGES AND BUG FIXES
The HUD class constructor now looks for an extra argument of the
model.scene instance. This is necessary to remove the extra rendering of
showing just the rotation arrows. Also with this we can enable autoclear
and disable all the manual depth checks. Now there is only one
renderer.render call with the hud showing the arrows without any fuss