Bug Fixes ---------
Fixed the transparency bug of materials with walls. This was caused by
walls that have corners supporting more than one room.
For each wall there is an interior and also an exterior wall.
Usually for the walls inside a house it is not necessary to draw the exterior
walls. As only the true boundary walls of a house needs to have both the
interior and exterior layers.